refactor : 도메인 영역
