+++ src/main/kotlin/racingcar/RacingView.kt
@@ -1,5 +1,7 @@
 package racingcar
 
+import racingcar.domain.Cars
+
 object InputView {
 
     private const val INPUT_CAR_NAMES = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분)."
