+++ src/main/kotlin/racingcar/domain/MoveStrategy.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 fun interface MoveStrategy {
     fun isMove(): Boolean
