+++ src/test/kotlin/racingcar/domain/CarSelectorTest.kt
@@ -1,8 +1,9 @@
-package racingcar
+package racingcar.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
+import racingcar.domain.CarSelector
 
 internal class CarSelectorTest {
