+++ src/test/kotlin/racingcar/domain/CarTest.kt
@@ -1,7 +1,8 @@
-package racingcar
+package racingcar.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import racingcar.domain.Car
 
 internal class CarTest {
