+++ src/test/kotlin/racingcar/domain/FailMoveStrategy.kt
@@ -1,4 +1,6 @@
-package racingcar
+package racingcar.domain
+
+import racingcar.domain.MoveStrategy
 
 class FailMoveStrategy : MoveStrategy {
     override fun isMove() = false
