+++ src/test/kotlin/racingcar/domain/SuccessMoveStrategy.kt
@@ -1,4 +1,6 @@
-package racingcar
+package racingcar.domain
+
+import racingcar.domain.MoveStrategy
 
 class SuccessMoveStrategy : MoveStrategy {
     override fun isMove() = true
